# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AERISYN – AI Weather Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /geocoding/search:
    get:
      operationId: searchLocations
      tags: [geocoding]
      summary: Search for locations by name
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeocodingResult"
  /geocoding/reverse:
    get:
      operationId: reverseGeocode
      tags: [geocoding]
      summary: Reverse geocode coordinates
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lon
          in: query
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Location info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeocodingResult"
...
  /analytics/historical-summary:
    get:
      operationId: getHistoricalSummary
      tags: [analytics]
      summary: Get historical weather summary statistics
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lon
          in: query
          required: true
          schema: { type: number }
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Historical summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoricalSummary"
  /analytics/comparison:
    post:
      operationId: compareCities
      tags: [analytics]
      summary: Compare weather across multiple cities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityComparisonInput"
      responses:
        "200":
          description: Comparison data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CityComparisonResult"
components:
  schemas:
...
    CurrentWeather:
      type: object
      required:
        - temperature
        - feels_like
        - humidity
        - pressure
        - wind_speed
        - wind_direction
        - visibility
        - cloud_cover
        - rainfall
        - uv_index
        - weather_code
        - weather_description
        - is_day
        - sunrise
        - sunset
      properties:
        temperature: { type: number }
        feels_like: { type: number }
        humidity: { type: number }
        pressure: { type: number }
        wind_speed: { type: number }
        wind_direction: { type: number }
        wind_gusts: { type: ["number", "null"] }
        visibility: { type: number }
        cloud_cover: { type: number }
        rainfall: { type: number }
        snowfall: { type: ["number", "null"] }
        uv_index: { type: number }
        weather_code: { type: integer }
        weather_description: { type: string }
        is_day: { type: integer }
        sunrise: { type: string }
        sunset: { type: string }
        moonrise: { type: ["string", "null"] }
        moonset: { type: ["string", "null"] }
...
    DailyWeather:
      type: object
      required: [date, temp_max, temp_min, weather_code, sunrise, sunset, precipitation_sum, precipitation_probability_max, wind_speed_max]
      properties:
        date: { type: string }
        temp_max: { type: number }
        temp_min: { type: number }
        weather_code: { type: integer }
        weather_description: { type: string }
        sunrise: { type: string }
        sunset: { type: string }
        uv_index_max: { type: ["number", "null"] }
        precipitation_sum: { type: number }
        precipitation_probability_max: { type: number }
        wind_speed_max: { type: number }
        wind_gusts_max: { type: ["number", "null"] }
...
    AirQuality:
      type: object
      required: [aqi, pm2_5, pm10, carbon_monoxide, nitrogen_dioxide, ozone, sulphur_dioxide, dust, uv_index]
      properties:
        aqi: { type: number }
        aqi_category: { type: string }
        pm2_5: { type: number }
        pm10: { type: number }
        carbon_monoxide: { type: number }
        nitrogen_dioxide: { type: number }
        ozone: { type: number }
        sulphur_dioxide: { type: number }
        dust: { type: ["number", "null"] }
        uv_index: { type: number }
        hourly:
          type: array
          items:
            $ref: "#/components/schemas/AirQualityHourly"
    AirQualityHourly:
      type: object
      required: [time, aqi, pm2_5, pm10]
      properties:
        time: { type: string }
        aqi: { type: number }
        pm2_5: { type: number }
        pm10: { type: number }
        ozone: { type: ["number", "null"] }
        nitrogen_dioxide: { type: ["number", "null"] }
    GeocodingResult:
      type: object
      required: [name, lat, lon, country]
      properties:
        id: { type: ["integer", "null"] }
        name: { type: string }
        lat: { type: number }
        lon: { type: number }
        country: { type: string }
        country_code: { type: ["string", "null"] }
        admin1: { type: ["string", "null"] }
        admin2: { type: ["string", "null"] }
        timezone: { type: ["string", "null"] }
        elevation: { type: ["number", "null"] }
        population: { type: ["integer", "null"] }
...
    HistoricalSummary:
      type: object
      required: [period, avg_temp, max_temp, min_temp, avg_humidity, total_rainfall, avg_wind_speed, avg_pressure, daily_data]
      properties:
        period: { type: string }
        avg_temp: { type: number }
        max_temp: { type: number }
        min_temp: { type: number }
        avg_humidity: { type: number }
        total_rainfall: { type: number }
        avg_wind_speed: { type: number }
        avg_pressure: { type: number }
        avg_aqi: { type: ["number", "null"] }
        daily_data:
          type: array
          items:
            $ref: "#/components/schemas/DailyWeather"
    CityComparisonInput:
      type: object
      required: [cities]
      properties:
        cities:
          type: array
          items:
            $ref: "#/components/schemas/CityLocation"
    CityLocation:
      type: object
      required: [name, lat, lon]
      properties:
        name: { type: string }
        lat: { type: number }
        lon: { type: number }
        country: { type: ["string", "null"] }
    CityComparisonResult:
      type: object
      required: [city, current_weather, air_quality]
      properties:
        city:
          $ref: "#/components/schemas/CityLocation"
        current_weather:
          $ref: "#/components/schemas/CurrentWeather"
        air_quality:
          $ref: "#/components/schemas/AirQuality"
        forecast_today:
          $ref: "#/components/schemas/DailyWeather"