# packages/api-zod/src/generated/types/geocodingResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AERISYN – AI Weather Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GeocodingResult {
  /** @nullable */
  id?: number | null;
  name: string;
  lat: number;
  lon: number;
  country: string;
  /** @nullable */
  country_code?: string | null;
  /** @nullable */
  admin1?: string | null;
  /** @nullable */
  admin2?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  elevation?: number | null;
  /** @nullable */
  population?: number | null;
}